+++ src/particle/src/engine/4C_particle_engine_container_bundle.hpp
@@ -76,7 +76,7 @@ namespace Particle
       FOUR_C_ASSERT(storedtypes_.contains(type), "container for particle type '{}' not stored!",
           enum_to_type_name(type));
 
-      return (containers_[type])[static_cast<int>(status)].get();
+      return (containers_[static_cast<int>(type)])[static_cast<int>(status)].get();
     };
 
     //! \name manipulate particle states of owned particles of specific type
@@ -96,7 +96,8 @@ namespace Particle
       FOUR_C_ASSERT(storedtypes_.contains(type), "container for particle type '{}' not stored!",
           enum_to_type_name(type));
 
-      ((containers_[type])[static_cast<int>(Status::Owned)])->scale_state(fac, state);
+      ((containers_[static_cast<int>(type)])[static_cast<int>(Status::Owned)])
+          ->scale_state(fac, state);
     };
 
     /*!
@@ -116,7 +117,7 @@ namespace Particle
       FOUR_C_ASSERT(storedtypes_.contains(type), "container for particle type '{}' not stored!",
           enum_to_type_name(type));
 
-      ((containers_[type])[static_cast<int>(Status::Owned)])
+      ((containers_[static_cast<int>(type)])[static_cast<int>(Status::Owned)])
           ->update_state(facA, stateA, facB, stateB);
     };
 
@@ -134,7 +135,8 @@ namespace Particle
       FOUR_C_ASSERT(storedtypes_.contains(type), "container for particle type '{}' not stored!",
           enum_to_type_name(type));
 
-      ((containers_[type])[static_cast<int>(Status::Owned)])->set_state(val, state);
+      ((containers_[static_cast<int>(type)])[static_cast<int>(Status::Owned)])
+          ->set_state(val, state);
     };
 
     /*!
@@ -149,7 +151,7 @@ namespace Particle
       FOUR_C_ASSERT(storedtypes_.contains(type), "container for particle type '{}' not stored!",
           enum_to_type_name(type));
 
-      ((containers_[type])[static_cast<int>(Status::Owned)])->clear_state(state);
+      ((containers_[static_cast<int>(type)])[static_cast<int>(Status::Owned)])->clear_state(state);
     };
 
     //! @}
@@ -167,7 +169,8 @@ namespace Particle
     inline void scale_state_all_containers(double fac, ParticleState state) const
     {
       for (const auto& type : storedtypes_)
-        ((containers_[type])[static_cast<int>(Status::Owned)])->scale_state(fac, state);
+        ((containers_[static_cast<int>(type)])[static_cast<int>(Status::Owned)])
+            ->scale_state(fac, state);
     };
 
     /*!
@@ -184,7 +187,7 @@ namespace Particle
         double facA, ParticleState stateA, double facB, ParticleState stateB) const
     {
       for (const auto& type : storedtypes_)
-        ((containers_[type])[static_cast<int>(Status::Owned)])
+        ((containers_[static_cast<int>(type)])[static_cast<int>(Status::Owned)])
             ->update_state(facA, stateA, facB, stateB);
     };
 
@@ -198,7 +201,8 @@ namespace Particle
     inline void set_state_all_containers(std::vector<double> val, ParticleState state) const
     {
       for (const auto& type : storedtypes_)
-        ((containers_[type])[static_cast<int>(Status::Owned)])->set_state(val, state);
+        ((containers_[static_cast<int>(type)])[static_cast<int>(Status::Owned)])
+            ->set_state(val, state);
     };
 
     /*!
@@ -210,7 +214,8 @@ namespace Particle
     inline void clear_state_all_containers(ParticleState state) const
     {
       for (const auto& type : storedtypes_)
-        ((containers_[type])[static_cast<int>(Status::Owned)])->clear_state(state);
+        ((containers_[static_cast<int>(type)])[static_cast<int>(Status::Owned)])
+            ->clear_state(state);
     };
 
     //! @}
@@ -228,7 +233,8 @@ namespace Particle
         ParticleStatus status) const
     {
       for (const auto& type : storedtypes_)
-        ((containers_[type])[static_cast<int>(status)])->check_and_decrease_container_size();
+        ((containers_[static_cast<int>(type)])[static_cast<int>(status)])
+            ->check_and_decrease_container_size();
     }
 
     /*!
@@ -240,7 +246,7 @@ namespace Particle
     inline void clear_all_containers_of_specific_status(ParticleStatus status) const
     {
       for (const auto& type : storedtypes_)
-        ((containers_[type])[static_cast<int>(status)])->clear_container();
+        ((containers_[static_cast<int>(type)])[static_cast<int>(status)])->clear_container();
     };
 
     //! @}
